Refactored rest.li equivalence classes.
